move functions into contrib, add tests
